requirements.txt
numpy
pyaudio
silero-vad
openai-whisper
pyttsx3
requests